Historicizing submission
I'm not sure why the changes to the index file are not showing.
I ran the index file through the w3. org validator and got several
errors relative to my use of the center element. It told me I should use
CSS, but I couldn't figure out how to make it apply to only the headings
on the first slide. The result looks like I want it despite the error.